Add FF Requset subscription and add multiple column lookup for attachment lookup
