+++ connect_transformations/lookup_ff_request/exceptions.py
@@ -1,0 +1,5 @@
+from connect_transformations.exceptions import BaseTransformationException
+
+
+class FFRequestLookupError(BaseTransformationException):
+    """Exception that happens when the FF request lookup fails"""
