+++ connect_transformations/static/transformations/airtable_lookup.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Airtable lookup</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/airtable_lookup.ae327bed82c601a2cd7c.js"></script><link href="../transformations/airtable_lookup.328186a96d00e5debdd4.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container"><div class="config-elem api-input"><label for="key-select">API Key</label> <input style="width: 86%;" id="key-input" placeholder="Enter an AirTable key"></div><div class="config-elem"><label for="base-select">AirTable Base</label> <select disabled="disabled" class="list" style="width: 80%;" id="base-select"></select></div><div class="config-elem"><label for="table-select">AirTable table name</label> <select disabled="disabled" class="list" style="width: 80%;" id="table-select"></select></div><div class="config-elem"><label for="table-select">Input column (Key)</label> <select disabled="disabled" class="list" style="width: 80%;" id="input-column-select"></select></div><div class="config-elem"><label for="table-select">AirTable Field Name (Key)</label> <select disabled="disabled" class="list" style="width: 80%;" id="field-select"></select></div></div><b>Columns Mapping</b><div id="content"></div><div class="button-container"><button disabled="disabled" id="add" class="button">ADD COLUMN</button></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Airtable lookup</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/airtable_lookup.77a65a38a2ecbe909dba.js"></script><link href="../transformations/airtable_lookup.328186a96d00e5debdd4.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container"><div class="config-elem api-input"><label for="key-select">API Key</label> <input style="width: 86%;" id="key-input" placeholder="Enter an AirTable key"></div><div class="config-elem"><label for="base-select">AirTable Base</label> <select disabled="disabled" class="list" style="width: 80%;" id="base-select"></select></div><div class="config-elem"><label for="table-select">AirTable table name</label> <select disabled="disabled" class="list" style="width: 80%;" id="table-select"></select></div><div class="config-elem"><label for="table-select">Input column (Key)</label> <select disabled="disabled" class="list" style="width: 80%;" id="input-column-select"></select></div><div class="config-elem"><label for="table-select">AirTable Field Name (Key)</label> <select disabled="disabled" class="list" style="width: 80%;" id="field-select"></select></div></div><b>Columns Mapping</b><div id="content"></div><div class="button-container"><button disabled="disabled" id="add" class="button">ADD COLUMN</button></div></div></body></html>
