+++ connect_transformations/static/transformations/attachment_lookup.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Attachment lookup</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/attachment_lookup.fadda5d1b92a1f248ff5.js"></script><link href="../transformations/attachment_lookup.7fff05b3722becd617b1.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div class="row"><div class="col"><div class="input-group"><label class="label" for="attachment">Attachment</label> <select materialize id="attachment"></select></div></div><div class="col"><div class="input-group"><label class="label" for="sheet">Sheet</label> <input materialize id="sheet"></div></div></div><div class="row"><div class="col"><div class="input-group"><label class="label" for="input-column">Input Column (Key)</label> <select materialize id="input-column"></select></div></div><div class="col"><div class="input-group"><label class="label" for="attachment-column">Attachment Field name (Key)</label> <input materialize id="attachment-column"></div></div></div><div id="mapping"><div class="subtitle">Columns mapping</div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Attachment lookup</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/attachment_lookup.1ed4bf0902783535fd5a.js"></script><link href="../transformations/attachment_lookup.7fff05b3722becd617b1.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="hidden"><div class="main-container" id="content"><div class="row"><div class="col"><div class="input-group"><label class="label" for="attachment">Attachment</label> <select materialize id="attachment"></select></div></div><div class="col"><div class="input-group"><label class="label" for="sheet">Sheet</label> <input materialize id="sheet"></div></div></div><div id="lookup"><div class="subtitle">Columns lookup</div></div><div class="button-container"><button id="add-lookup-button" class="button">ADD COLUMN</button></div><div id="mapping"><div class="subtitle">Columns mapping</div></div></div></div></body></html>
