+++ connect_transformations/static/transformations/copy.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Copy</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/copy.ccb409b9c25d42e5256a.js"></script><link href="../transformations/copy.03647bd51a5a3976242d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="application"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Copy</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/copy.1e45e8962a0d1988b7a1.js"></script><link href="../transformations/copy.03647bd51a5a3976242d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app" class="application"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
