+++ connect_transformations/static/transformations/currency_conversion.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency conversion</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/currency_conversion.e74c4f8cf1bec6f654cb.js"></script><link href="../transformations/currency_conversion.03647bd51a5a3976242d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Currency conversion</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/currency_conversion.b58a42cd419e168f7094.js"></script><link href="../transformations/currency_conversion.03647bd51a5a3976242d.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD COLUMN</button></div></div></body></html>
