+++ connect_transformations/static/transformations/filter_row.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Filter row</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/filter_row.d34d309272848e471254.js"></script><link href="../transformations/filter_row.d48f57a82ff30cea5a02.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="value">Value</label> <input id="value" class="input" placeholder="Value"/></div></div><div class="button-container"><button id="add" class="button">ADD VALUE</button></div><div class="input-group"><p class="radio-buttons">Choose how to filter rows:</p><div class="radio-buttons"><input type="radio" id="match" name="filter_condition" value="match"/> <label class="label" for="match">By match</label> <input type="radio" id="mismatch" name="filter_condition" value="mismatch"/> <label class="label" for="mismatch">By mismatch</label></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Filter row</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/filter_row.3262aa5f35c195600137.js"></script><link href="../transformations/filter_row.d48f57a82ff30cea5a02.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="value">Value</label> <input id="value" class="input" placeholder="Value"/></div></div><div class="button-container"><button id="add" class="button">ADD VALUE</button></div><div class="input-group"><p class="radio-buttons">Choose how to filter rows:</p><div class="radio-buttons"><input type="radio" id="match" name="filter_condition" value="match"/> <label class="label" for="match">By match</label> <input type="radio" id="mismatch" name="filter_condition" value="mismatch"/> <label class="label" for="mismatch">By mismatch</label></div></div></div></body></html>
