+++ connect_transformations/static/transformations/formula.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"/><title>Transformations/Formula</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/formula.adb97ac61884100f30f2.js"></script><link href="../transformations/formula.425c1b7b3e6c057e75d2.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></body></html>
+<!doctype html><html lang="en"><head><meta charset="utf-8"/><title>Transformations/Formula</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/formula.ed8590437b8cd3c5851f.js"></script><link href="../transformations/formula.425c1b7b3e6c057e75d2.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container" id="content"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></body></html>
