+++ connect_transformations/static/transformations/lookup_product_item.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup product item</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_product_item.13b6a0d4254ac4503f19.js"></script><link href="../transformations/lookup_product_item.421df6826da3f2351526.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include the CloudBlue item ID and MPN.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="subtitle no-product" id="no_product">Warning: No product associated with this data stream was found. Attach a product to the stream on the stream Configuration tab, enter the Product ID below, or indicate a column from which the Product ID value will be taken.</div><div class="input-group" id="product_id_radio_group" style="margin-bottom: 8px"><p class="radio-buttons">Find product by:</p><div class="radio-buttons"><input type="radio" id="by_product_id" name="product_id_radio" value="product_id"/> <label class="label" for="by_product_id">Product ID</label><br/><input type="radio" id="by_product_column" name="product_id_radio" value="product_column"/> <label class="label" for="by_product_column">Column containing product ID</label></div></div><div class="input-group" id="product_id_input"><label class="label" for="product_id">Product ID</label> <input class="input" id="product_id"/></div><div class="input-group" id="product_column_input"><label class="label" for="product_id_column">Product ID Column</label> <select class="input" id="product_id_column"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">Choose what to do if a product item is not found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave the row empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Transformation will fail</label></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup product item</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_product_item.4d370f09875d97641195.js"></script><link href="../transformations/lookup_product_item.421df6826da3f2351526.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include the CloudBlue item ID and MPN.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="subtitle no-product" id="no_product">Warning: No product associated with this data stream was found. Attach a product to the stream on the stream Configuration tab, enter the Product ID below, or indicate a column from which the Product ID value will be taken.</div><div class="input-group" id="product_id_radio_group" style="margin-bottom: 8px"><p class="radio-buttons">Find product by:</p><div class="radio-buttons"><input type="radio" id="by_product_id" name="product_id_radio" value="product_id"/> <label class="label" for="by_product_id">Product ID</label><br/><input type="radio" id="by_product_column" name="product_id_radio" value="product_column"/> <label class="label" for="by_product_column">Column containing product ID</label></div></div><div class="input-group" id="product_id_input"><label class="label" for="product_id">Product ID</label> <input class="input" id="product_id"/></div><div class="input-group" id="product_column_input"><label class="label" for="product_id_column">Product ID Column</label> <select class="input" id="product_id_column"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><label class="label" for="prefix">Prefix</label> <input class="input" id="prefix" placeholder="Prefix"/></div><div class="input-group"><p class="radio-buttons">Choose what to do if a product item is not found:</p><div class="radio-buttons"><input type="radio" id="leave_empty" name="if_not_found" value="leave_empty"/> <label class="label" for="leave_empty">Leave the row empty</label> <input type="radio" id="fail" name="if_not_found" value="Fail"/> <label class="label" for="fail">Transformation will fail</label></div></div></div></div></body></html>
