+++ connect_transformations/static/transformations/lookup_subscription.html
@@ -1,1 +1,1 @@
-<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup subscription</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_subscription.b72713276a2a36595191.js"></script><link href="../transformations/lookup_subscription.421df6826da3f2351526.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><p class="radio-buttons label">Choose what to do if a subscription is not found:</p><div class="radio-buttons input"><input type="radio" id="not_found_leave_empty" name="if_not_found" value="leave_empty"/> <label for="not_found_leave_empty">Leave empty</label> <input type="radio" id="not_found_fail" name="if_not_found" value="fail"/> <label for="not_found_fail">Fail</label></div></div><div class="input-group"><p class="radio-buttons label">Choose what to do if multiple subscriptions are found:</p><div class="radio-buttons input"><input type="radio" id="multiple_use_most_actual" name="if_multiple" value="use_most_actual"/> <label for="multiple_use_most_actual">Use the most recent</label> <input type="radio" id="multiple_leave_empty" name="if_multiple" value="leave_empty"/> <label for="multiple_leave_empty">Leave empty</label> <input type="radio" id="multiple_fail" name="if_multiple" value="fail"/> <label for="multiple_fail">Fail</label></div></div><div class="output"><label>Output columns:</label><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></div></div></body></html>
+<!doctype html><html><head><meta charset="utf-8"/><title>Transformations/Lookup subscription</title><script defer="defer" src="../vendors.b7829ba6a3fa12b6e5cb.js"></script><script defer="defer" src="../transformations/lookup_subscription.476596bbddd038d638d2.js"></script><link href="../transformations/lookup_subscription.421df6826da3f2351526.css" rel="stylesheet"></head><body><div id="loader"></div><div id="app"><div class="main-container lookup" id="content"><div class="subtitle">Supported search criteria include Subscription External ID, ID and Parameter values.</div><div class="input-group"><label class="label" for="criteria">Criterion</label> <select class="input" id="criteria"></select></div><div class="input-group" id="param_name_group"><label class="label" for="parameter">Parameter Name</label> <select class="input" id="parameter"></select></div><div class="input-group"><label class="label" for="column">Input Column</label> <select class="input" id="column"></select></div><div class="input-group"><p class="radio-buttons label">Choose what to do if a subscription is not found:</p><div class="radio-buttons input"><input type="radio" id="not_found_leave_empty" name="if_not_found" value="leave_empty"/> <label for="not_found_leave_empty">Leave empty</label> <input type="radio" id="not_found_fail" name="if_not_found" value="fail"/> <label for="not_found_fail">Fail</label></div></div><div class="input-group"><p class="radio-buttons label">Choose what to do if multiple subscriptions are found:</p><div class="radio-buttons input"><input type="radio" id="multiple_use_most_actual" name="if_multiple" value="use_most_actual"/> <label for="multiple_use_most_actual">Use the most recent</label> <input type="radio" id="multiple_leave_empty" name="if_multiple" value="leave_empty"/> <label for="multiple_leave_empty">Leave empty</label> <input type="radio" id="multiple_fail" name="if_multiple" value="fail"/> <label for="multiple_fail">Fail</label></div></div><div class="output"><label>Output columns:</label><div id="output-columns"></div><div class="button-container"><button id="add" class="button">ADD</button></div></div></div></div></body></html>
