+++ connect_transformations/transformations.py
@@ -15,6 +15,7 @@ from connect_transformations.copy_columns.mixins import CopyColumnTransformation
 from connect_transformations.currency_conversion.mixins import CurrencyConverterTransformationMixin
 from connect_transformations.filter_row.mixins import FilterRowTransformationMixin
 from connect_transformations.formula.mixins import FormulaTransformationMixin
+from connect_transformations.lookup_ff_request.mixins import LookupFFRequestTransformationMixin
 from connect_transformations.lookup_product_items.mixins import (
     LookupProductItemsTransformationMixin,
 )
@@ -33,6 +34,7 @@ class StandardTransformationsApplication(
     CurrencyConverterTransformationMixin,
     FilterRowTransformationMixin,
     FormulaTransformationMixin,
+    LookupFFRequestTransformationMixin,
     LookupSubscriptionTransformationMixin,
     SplitColumnTransformationMixin,
     LookupProductItemsTransformationMixin,
