+++ ui/pages/transformations/attachment_lookup.html
@@ -30,19 +30,13 @@ All rights reserved.
           </div>
         </div>
       </div>
-      <div class="row">
-        <div class="col">
-          <div class="input-group">
-            <label class="label" for="input-column">Input Column (Key)</label>
-            <select materialize id="input-column"></select>
-          </div>
+      <div id="lookup">
+        <div class="subtitle">
+          Columns lookup
         </div>
-        <div class="col">
-          <div class="input-group">
-            <label class="label" for="attachment-column">Attachment Field name (Key)</label>
-            <input materialize type="text" id="attachment-column"></select>
-          </div>
-        </div>
+      </div>
+      <div class="button-container">
+        <button id="add-lookup-button" class="button">ADD COLUMN</button>
       </div>
       <div id="mapping">
         <div class="subtitle">
