+++ ui/pages/transformations/lookup_ff_request.html
@@ -1,0 +1,87 @@
+<!--
+Copyright (c) 2023, CloudBlue LLC
+All rights reserved.
+-->
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="utf-8" />
+    <title>
+        <%= htmlWebpackPlugin.options.title %>
+    </title>
+</head>
+
+<body>
+    <div id="loader"></div>
+    <div id="app">
+        <div class="main-container lookup" id="content">
+            <div class="subtitle">
+                Supported search by parameter and optional
+                search by Subscription ID or Subscription External ID.
+                Result contains only one item information, matching selected criteria.
+            </div>
+
+            <div class="input-group">
+                <p>Choose search by Subscription ID</p>
+                <label class="label" for="criteria">Subscription ID type</label>
+                <select class="input" id="criteria"></select>
+            </div>
+            <div class="input-group" id="subscription_column">
+                <label class="label" for="column">Input Column</label>
+                <select class="input" id="column"></select>
+            </div>
+
+            <div class="input-group">
+                <p>Choose search by parameter</p>
+                <label class="label" for="parameter">Parameter name</label>
+                <select class="input" id="parameter"></select>
+            </div>
+            <div>
+                <label class="label" for="parameter_column">Input Column</label>
+                <select class="input" id="parameter_column"></select>
+            </div>
+
+            <div class="input-group">
+                <p>Choose Item match criteria</p>
+                <label class="label" for="item">Item ID type</label>
+                <select class="input" id="item"></select>
+            </div>
+            <div id="item_column_group">
+                <label class="label" for="item_column">Input Column</label>
+                <select class="input" id="item_column"></select>
+            </div>
+
+            <div class="input-group">
+                <p class="radio-buttons label">Choose what to do if match is not found:</p>
+                <div class="radio-buttons input">
+                    <input type="radio" id="not_found_leave_empty" name="if_not_found" value="leave_empty" />
+                    <label for="not_found_leave_empty">Leave empty</label>
+                    <input type="radio" id="not_found_fail" name="if_not_found" value="fail" />
+                    <label for="not_found_fail">Fail</label>
+                </div>
+            </div>
+            <div class="input-group">
+                <p class="radio-buttons label">Choose what to do if multiple matches are found:</p>
+                <div class="radio-buttons input">
+                    <input type="radio" id="multiple_use_most_actual" name="if_multiple" value="use_most_actual" />
+                    <label for="multiple_use_most_actual">Use the most recent</label>
+                    <input type="radio" id="multiple_leave_empty" name="if_multiple" value="leave_empty" />
+                    <label for="multiple_leave_empty">Leave empty</label>
+                    <input type="radio" id="multiple_fail" name="if_multiple" value="fail" />
+                    <label for="multiple_fail">Fail</label>
+                </div>
+            </div>
+            <div class="output">
+                <label>Output columns:</label>
+                <div id="output-columns">
+                </div>
+                <div class="button-container">
+                    <button id="add" class="button">ADD</button>
+                </div>
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>
