+++ ui/src/pages/transformations/attachment_lookup.js
@@ -13,6 +13,7 @@ import {
   hideComponent,
   hideError,
   showComponent,
+  showError,
 } from '../../components';
 
 import {
@@ -47,6 +48,45 @@ export const fillSelect = (options, id, value) => {
     select.appendChild(option);
   });
 };
+
+
+export const createLookupRow = (parent, index, options, input, output) => {
+  const item = document.createElement('div');
+  item.classList.add('list-wrapper');
+  item.id = `wrapper-${index}`;
+  item.innerHTML = `
+      <select class="list" style="width: 35%;" ${input ? `value="${input.id}"` : ''}>
+        ${options.map((column) => `
+          <option value="${column.id}" ${input && input.id === column.id ? 'selected' : ''}>
+            ${getColumnLabel(column)}
+          </option>`).join(' ')}
+      </select>
+      <input type="text" placeholder="Attachment Field name" style="width: 35%;" ${output ? `value="${output}"` : ''} />
+      <button id="delete-lk-${index}" style="display: initial; width: 20%" class="button delete-button">DELETE</button>
+    `;
+  parent.appendChild(item);
+
+  document.getElementById(`delete-lk-${index}`).addEventListener('click', () => {
+    if (document.getElementsByClassName('list-wrapper').length === 1) {
+      showError('You need to have at least one lookup');
+    } else {
+      document.getElementById(`wrapper-${index}`).remove();
+      const buttons = document.getElementsByClassName('delete-button');
+      if (buttons.length === 1) {
+        buttons[0].disabled = true;
+      }
+    }
+  });
+  const buttons = document.getElementsByClassName('delete-button');
+  for (let i = 0; i < buttons.length; i += 1) {
+    if (buttons.length === 1) {
+      buttons[i].disabled = true;
+    } else {
+      buttons[i].disabled = false;
+    }
+  }
+};
+
 
 export const createMappingRow = (index, from, to) => {
   let lastRowIndex = 0;
@@ -94,11 +134,12 @@ export const createMappingRow = (index, from, to) => {
   });
 };
 
-export const loockupSpreadsheet = (app) => {
+export const lookupSpreadsheet = (app) => {
   if (!app) return;
 
   let attachments = [];
   let columns = [];
+  let rowIndex = 0;
 
   app.listen('config', async (config) => {
     try {
@@ -107,25 +148,30 @@ export const loockupSpreadsheet = (app) => {
           stream: { id: streamId },
           available_columns: availableColumns,
         },
+        columns: { input: inputColumns },
         settings,
       } = config;
 
       attachments = await getAttachments(streamId);
       columns = availableColumns;
 
+      const content = document.getElementById('lookup');
+
       if (settings) {
         const {
           file,
           sheet,
-          map_by: {
-            input_column: inputColumnName,
-            attachment_column: attachmentColumn,
-          },
+          map_by: mapBy,
           mapping,
         } = settings;
 
-        const inputColumn = columns.find((item) => item.name === inputColumnName);
-        fillSelect(columns, 'input-column', inputColumn.id);
+        mapBy.forEach((item, i) => {
+          const inputColumn = inputColumns.find((col) => col.name === item.input_column);
+          const outputColumn = item.attachment_column;
+          rowIndex = i;
+          createLookupRow(content, rowIndex, columns, inputColumn, outputColumn);
+        });
+
         const attachmentFound = attachments.find((item) => item.file === file);
         let fileId = null;
         if (attachmentFound == null) {
@@ -135,16 +181,20 @@ export const loockupSpreadsheet = (app) => {
           fileId = attachmentFound.id;
         }
         fillSelect(attachments, 'attachment', fileId);
-        document.getElementById('attachment-column').value = attachmentColumn;
         document.getElementById('sheet').value = sheet;
         mapping.forEach((item, index) => {
           createMappingRow(index, item.from, item.to);
         });
       } else {
-        fillSelect(columns, 'input-column');
         fillSelect(attachments, 'attachment');
         createMappingRow(0);
+        createLookupRow(content, rowIndex, columns);
       }
+
+      document.getElementById('add-lookup-button').addEventListener('click', () => {
+        rowIndex += 1;
+        createLookupRow(content, rowIndex, columns);
+      });
     } catch (error) {
       app.emit('validation-error', error);
     } finally {
@@ -160,9 +210,21 @@ export const loockupSpreadsheet = (app) => {
       const fileId = getRequiredValue('attachment', 'Please select attachment');
       const { file } = attachments.find((item) => item.id === fileId);
       const sheet = document.getElementById('sheet').value;
-      const inputColumnId = getRequiredValue('input-column', 'Please select input column');
-      const inputColumn = columns.find((item) => item.id === inputColumnId);
-      const attachmentColumn = getRequiredValue('attachment-column', 'Please select attachment column');
+
+      const inputColumns = [];
+      const mapBy = [];
+      const lookups = document.querySelectorAll('#lookup .list-wrapper');
+      lookups.forEach((lookup) => {
+        const inputId = lookup.children[0].value;
+        const attchName = lookup.children[1].value;
+        const inputColumn = columns.find((column) => column.id === inputId);
+        if (inputId && attchName) {
+          mapBy.push({ input_column: inputColumn.name, attachment_column: attchName });
+          inputColumns.push(inputColumn);
+        } else {
+          throw new Error('Please fill all mapping rows');
+        }
+      });
 
       const outputColumns = [];
       const mapping = [];
@@ -184,14 +246,11 @@ export const loockupSpreadsheet = (app) => {
         settings: {
           file,
           sheet,
-          map_by: {
-            input_column: inputColumn.name,
-            attachment_column: attachmentColumn,
-          },
+          map_by: mapBy,
           mapping,
         },
         columns: {
-          input: [inputColumn],
+          input: inputColumns,
           output: outputColumns,
         },
       };
@@ -207,4 +266,4 @@ export const loockupSpreadsheet = (app) => {
   });
 };
 
-createApp({ }).then(loockupSpreadsheet);
+createApp({ }).then(lookupSpreadsheet);
