+++ webpack.config.js
@@ -33,6 +33,7 @@ const pages = populateHtmlPlugins([
   "Transformations/Copy",
   "Transformations/Attachment lookup",
   "Transformations/Lookup subscription",
+  "Transformations/Lookup FF request",
   "Transformations/Lookup product item",
   "Transformations/Currency conversion",
   "Transformations/Manual",
@@ -51,6 +52,7 @@ module.exports = {
     ['transformations/filter_row']: __dirname + "/ui/src/pages/transformations/filter_row.js",
     ['transformations/formula']: __dirname + "/ui/src/pages/transformations/formula.js",
     ['transformations/lookup_subscription']: __dirname + "/ui/src/pages/transformations/lookup_subscription.js",
+    ['transformations/lookup_ff_request']: __dirname + "/ui/src/pages/transformations/lookup_ff_request.js",
     ['transformations/lookup_product_item']: __dirname + "/ui/src/pages/transformations/lookup_product_item.js",
     ['transformations/manual']: __dirname + "/ui/src/pages/transformations/manual.js",
     ['transformations/split_columns']: __dirname + "/ui/src/pages/transformations/split_columns.js",
